refactor(components): reorganize Timer and WalletConnectButton directory structure
- Move Timer from src/ui to components/Main/Timer/
- Move WalletConnectButton from src/ui to components/header/
- Update imports in page.tsx and Header.tsx

As per PR review feedback to better organize functional components
