+++ src/components/Main/Timer/Timer.tsx
